[test] modify getApi42.js to useApi42.js and move out getUriPart, getUserData method from useApi42 object.
